Remove dependency for parserUtil
AB3 parserUtil still exists.

Let's remove it.

ParserUtil is replaced by SocheduleParserUtil officially. Test methods have also been updated.
