+++ src/main/java/seedu/address/logic/parser/DeleteTaskCommandParser.java
@@ -18,7 +18,7 @@ public class DeleteTaskCommandParser implements Parser<DeleteTaskCommand> {
      */
     public DeleteTaskCommand parse(String args) throws ParseException {
         try {
-            Index index = ParserUtil.parseIndex(args);
+            Index index = SocheduleParserUtil.parseIndex(args);
             return new DeleteTaskCommand(index);
         } catch (ParseException pe) {
             throw new ParseException(
